backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 异步数据库
asyncpg>=0.29.0

# Redis
redis>=5.0.0

# AI/LLM
openai>=1.10.0
langgraph>=0.0.40
langchain>=0.1.0

# WebSocket
websockets>=12.0
sse-starlette>=2.0.0

# 工具
python-dotenv>=1.0.0
python-multipart>=0.0.6

# 开发工具
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
